requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
google-generativeai==0.3.2
python-dotenv==1.0.0
selenium==4.15.2
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0 